backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0